+++ src/containers/Counter/Counter.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 
 import CounterControl from '../../components/CounterControl/CounterControl';
 import CounterOutput from '../../components/CounterOutput/CounterOutput';
-import * as actionCreators from '../../store/actions/actions.js';
+import * as actionCreators from '../../store/actions/index';
 
 class Counter extends Component {
     render () {
