+++ src/store/actions/index.js
@@ -1,0 +1,11 @@
+export {
+  add,
+  substract,
+  increment,
+  decrement
+} from './counter';
+
+export {
+  storeResult,
+  deleteResult
+} from './result';
