+++ src/store/reducers/result.js
@@ -1,4 +1,4 @@
-import * as actionTypes from '../actions/actions';
+import * as actionTypes from '../actions/actionTypes';
 
 const initialState = {
     results: []
@@ -9,7 +9,7 @@ const reducer = ( state = initialState, action ) => {
         case actionTypes.STORE_RESULT:
             return {
                 ...state,
-                results: state.results.concat({id: new Date(), value: action.result})
+                results: state.results.concat({id: new Date(), value: action.result * 2})
             }
         case actionTypes.DELETE_RESULT:
             // const id = 2;
